Explicitly use new default max length for `EmailField` per Django >= 1.8
From Django version 1.8 the default `max_length` for `EmailField` is 254
characters instead of the previous 75 character limit. This change was
made in Django to "be compliant with RFCs 3696 and 5321".

Setting this length explicitly means the new 0002 migration from
upstream will make sense for Django 1.7 projects, where the new-style
migration will get applied (instead of South migrations) but the
Django default max length for `EmailField` is still 75 characters.